версия задания 8.3 для проверки
