+++ css/style.css
@@ -293,9 +293,9 @@ fieldset {
   -webkit-box-align: center;
       -ms-flex-align: center;
           align-items: center;
+  flex-shrink: 0;
 
   margin-right: 10px;
-  padding: 0 15px;
 
   width: 70px;
   height: 70px;
@@ -304,6 +304,11 @@ fieldset {
 
   border-radius: 5px;
   background-color: #dadada;
+  overflow: hidden;
+}
+
+.ad-form-header__preview img {
+  object-fit: cover;
 }
 
 .ad-form__field {
@@ -548,6 +553,7 @@ fieldset {
   width: 100%;
   height: 70px;
   padding: 20px 10px 5px;
+  margin-bottom: 10px;
 
   text-align: center;
   font-size: 14px;
@@ -571,14 +577,20 @@ fieldset {
 }
 
 .ad-form__photo {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 5px;
   margin-right: 10px;
   margin-bottom: 10px;
 
-  width: 70px;
-  height: 70px;
+  min-width: 70px;
+  max-width: 370px;
+  min-height: 70px;
 
   background-color: #e4e4de;
   border-radius: 5px;
+  overflow: hidden;
+
 }
 
 .ad-form__element--submit {
